parser: Make the parser generic to allow custom pseudo-elements and pseudo-classes
This doesn't make the crate to compile, but just wanted to check with
Simon if this is the approach he was expecting.

This leaves out of rust-selectors all non tree-structural
pseudo-classes, and leverages pseudo-element parsing to the associated
trait.

I'm bad at naming things, so any more idiomatic name that SelectorImpl
would be appreciated.